move constants to their own file, slight refactors to match Gameboy design
