+++ js/constants.js
@@ -1,0 +1,3 @@
+export const NUM_ROWS = 18;
+export const NUM_COLS = 10;
+export const GRID_SPACE = 20;
